+++ src/main/java/com/festival/domain/foodTruck/data/dto/request/FoodTruckRequest.java
@@ -29,10 +29,10 @@ public class FoodTruckRequest implements Serializable {
     private List<MultipartFile> subFiles;
 
     @NotNull(message = "위치를 입력해주세요.")
-    private int latitude;
+    private float latitude;
 
     @NotNull(message = "위치를 입력해주세요.")
-    private int longitude;
+    private float longitude;
 
     @NotNull(message = "상태를 입력해주세요.")
     private Boolean state;
