+++ src/main/java/com/festival/domain/foodTruck/data/dto/response/FoodTruckResponse.java
@@ -20,15 +20,15 @@ public class FoodTruckResponse {
     private String mainFilePath;
     private List<String> subFilePaths;
 
-    private int latitude;
-    private int longitude;
+    private float latitude;
+    private float longitude;
 
     private Boolean state;
 
     @Builder
     @QueryProjection
     public FoodTruckResponse(String title, String subTitle, String content,
-                       String mainFilePath, List<String> subFilePaths, int latitude, int longitude, Boolean state) {
+                       String mainFilePath, List<String> subFilePaths, float latitude, float longitude, Boolean state) {
         this.title = title;
         this.subTitle = subTitle;
         this.content = content;
