+++ src/main/java/com/festival/domain/info/festivalEvent/data/dto/FestivalEventReq.java
@@ -20,14 +20,14 @@ public class FestivalEventReq {
     private String content;
 
     @NotNull(message = "위치를 입력해주세요.")
-    private int latitude;
+    private float latitude;
 
     @NotNull(message = "위치를 입력해주세요.")
-    private int longitude;
+    private float longitude;
 
     @NotNull(message = "상태를 입력해주세요.")
     private Boolean state;
-    public FestivalEventReq(String title, String subTitle, String content, MultipartFile mainFile, List<MultipartFile> subFiles, int latitude, int longitude, Boolean state) {
+    public FestivalEventReq(String title, String subTitle, String content, MultipartFile mainFile, List<MultipartFile> subFiles, float latitude, float longitude, Boolean state) {
         this.title = title;
         this.subTitle = subTitle;
         this.content = content;
